+++ CHANGELOG.md
@@ -6,6 +6,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [x.x.x]
 ### Added
+- Display number of filtered variants vs number of total variants in variants page
 - Search case by HPO terms
 
 ### Fixed